Split out getPlansBySite(Id)
